% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_ats_iridium.R
\name{scrape_ats_iridium}
\alias{scrape_ats_iridium}
\title{Scrape data from ATS website}
\usage{
scrape_ats_iridium(bttn_nm = NULL, usr = NULL, pwd = NULL)
}
\arguments{
\item{bttn_nm}{The name of the button to download data from}

\item{usr}{The username associated with the account}

\item{pwd}{The password used with the supplied username}
}
\value{
A tibble with automatic parsing of column types
}
\description{
Scrape data from ATS website
}
\details{
The url where the function attempts to gain access is atsdat.net.

To download from a new button name use the developer tools in your browser and record the name of the button as described there.  For example, the download all button has CSS name 'ctl00$ContentPlaceHolder1$DownloadAll3'.
}
\examples{
scrape_ats(usr = "my_username", pwd = "secret_code")
}
