% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_celltrack.R
\name{scrape_celltrack}
\alias{scrape_celltrack}
\title{Cell Track web scraping utilities}
\usage{
scrape_celltrack(usr, pwd, page_count = 1)
}
\arguments{
\item{usr}{The username associated with your account}

\item{pwd}{User's password}

\item{page_count}{The number of pages displayed in the Unit List or NULL, when NULL an attempt is made to detect the number of pages}
}
\value{
A tibble
}
\description{
Cell Track web scraping utilities
}
\examples{
scrape_celltrack("my_name", "my_secret", NULL)

}
